+++ CHANGELOG.md
@@ -11,6 +11,7 @@ adheres to [Semantic Versioning](https://semver.org/).
 
 ### :house: Internal
 
+- Necessary code changes following dev dependency update: pytest
 - Add tests for CPython 3.11
 - Use CPython 3.11 for misc. tests