+++ CHANGELOG.md
@@ -5,9 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project
 adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.9.0] - 2022-11-06
 
-[unreleased]: https://github.com/rogdham/bigxml/compare/v0.8.0...HEAD
+[0.9.0]: https://github.com/rogdham/bigxml/compare/v0.8.0...v0.9.0
+
+v0.9.0 introduces custom exceptions.
 
 ### :boom: Breaking changes